+++ src/cjs/geomutils.d.ts
@@ -24,5 +24,13 @@ export declare const geomutils: {
      * @return {Line[]} An array of n-1 lines secting the given angle in point A into n equal sized angle sections. The lines' first vertex is A.
      */
     nsectAngle(pA: Vertex, pB: Vertex, pC: Vertex, n: number): Array<Line>;
+    /**
+     * Wrap the value (e.g. an angle) into the given range of [0,max).
+     *
+     * @name wrapMax
+     * @param {number} x - The value to wrap.
+     * @param {number} max - The max bound to use for the range.
+     * @return {number} The wrapped value inside the range [0,max).
+     */
     wrapMax(x: number, max: number): number;
 };