+++ src/ts/VEllipse.ts
@@ -7,7 +7,8 @@
  * @modified 2021-02-14 Added functions `radiusH` and `radiusV`.
  * @modified 2021-02-26 Added helper function `decribeSVGArc(...)`.
  * @modified 2021-03-01 Added attribute `rotation` to allow rotation of ellipses.
- * @version  1.2.1
+ * @modified 2021-03-04 Added the `vertAt` and `perimeter` functions.
+ * @version  1.2.2
  *
  * @file VEllipse
  * @fileoverview Ellipses with a center and an x- and a y-axis (stored as a vertex).
@@ -77,7 +78,7 @@ export class VEllipse implements SVGSerializable {
      * @param {Vertex} rotation - [optional, default=0] The rotation of this ellipse.
      * @name VEllipse
      **/
-    constructor( center:Vertex, axis:Vertex, rotation?:number ) {
+    constructor( center:Vertex, axis:Vertex, rotation?:number) {
 	this.uid = UIDGenerator.next();
 	this.center = center;
 	this.axis = axis;
@@ -93,9 +94,6 @@ export class VEllipse implements SVGSerializable {
      * @return {number} The unsigned horizontal radius of this ellipse.
      */
     radiusH() : number {
-	// return Math.abs(this.axis.x - this.center.x);
-	// Rotate axis back to origin before calculating radius
-	// return Math.abs(new Vertex(this.axis).rotate(-this.rotation,this.center).x - this.center.x);
 	return Math.abs( this.signedRadiusH() );
     };
 
@@ -123,9 +121,6 @@ export class VEllipse implements SVGSerializable {
      * @return {number} The unsigned vertical radius of this ellipse.
      */
     radiusV() : number {
-	// return Math.abs(this.axis.y - this.center.y);
-	// Rotate axis back to origin before calculating radius
-	// return Math.abs(new Vertex(this.axis).rotate(-this.rotation,this.center).y - this.center.y);
 	return Math.abs( this.signedRadiusV() );
     };
 
@@ -144,13 +139,41 @@ export class VEllipse implements SVGSerializable {
 	return new Vertex(this.axis).rotate(-this.rotation,this.center).y - this.center.y;
     };
 
-    
+
+    /**
+     * Get the vertex on the ellipse's outline at the given angle.
+     *
+     * @method vertAt
+     * @instance
+     * @memberof VEllipse
+     * @param {number} angle - The angle to determine the vertex at.
+     * @return {Vertex} The vertex on the outline at the given angle.
+     */
     vertAt( angle:number ) : Vertex {
 	// Tanks to Narasinham for the vertex-on-ellipse equations
 	// https://math.stackexchange.com/questions/22064/calculating-a-point-that-lies-on-an-ellipse-given-an-angle
 	const a : number = this.radiusH();
 	const b : number = this.radiusV();
 	return new Vertex( VEllipse.utils.polarToCartesian( this.center.x, this.center.y, a, b, angle) ).rotate( this.rotation, this.center );
+    };
+
+
+    /**
+     * Get the perimeter of this ellipse.
+     *
+     * @method perimeter
+     * @instance
+     * @memberof VEllipse
+     * @return {number}
+     */
+    perimeter() : number {
+	// This method does not use an iterative approximation to determine the perimeter, but it uses
+	// a wonderful closed approximation found by Srinivasa Ramanujan.
+	// Matt Parker made a neat video about it:
+	//    https://www.youtube.com/watch?v=5nW3nJhBHL0
+	const a : number = this.radiusH();
+	const b : number = this.radiusV();
+	return Math.PI * ( 3*(a+b) - Math.sqrt( (3*a+b)*(a+3*b) ) );
     };