+++ src/ts/VEllipseSector.ts
@@ -11,9 +11,10 @@
 // TODO: add to PlotBoilerplate.add(...)
 
 
-import { VEllipse } from "./VEllipse";
+import { geomutils } from "./geomutils";
 import { SVGPathParams, UID, XYCoords } from "./interfaces";
 import { UIDGenerator } from "./UIDGenerator";
+import { VEllipse } from "./VEllipse";
 import { Vertex } from "./Vertex";
 
 /**
@@ -87,15 +88,15 @@ export class VEllipseSector {
 	this.endAngle = endAngle;
     }
     
+    
     static ellipseSectorUtils = {
 	/**
-	 * Helper function to convert a circle section as SVG arc params (for the `d` attribute).
-	 * Found at: https://stackoverflow.com/questions/5736398/how-to-calculate-the-svg-path-for-an-arc-of-a-circle
-	 *
-	 * TODO: generalize for ellipses (two radii).
+	 * Helper function to convert an elliptic section to SVG arc params (for the `d` attribute).
+	 * Inspiration found at: 
+	 *    https://stackoverflow.com/questions/5736398/how-to-calculate-the-svg-path-for-an-arc-of-a-circle
 	 *
 	 * @param {boolean} options.moveToStart - If false (default=true) the initial 'Move' command will not be used.
-	 * @return [ 'A', radiusH, radiusV, rotation=0, largeArcFlag=1|0, sweepFlag=0, endx, endy ]
+	 * @return [ 'A', radiusH, radiusV, rotation, largeArcFlag=1|0, sweepFlag=0, endx, endy ]
 	 */
 	describeSVGArc : ( x : number, y : number,
 			   radiusH : number,
@@ -111,33 +112,28 @@ export class VEllipseSector {
 	    if( typeof rotation === 'undefined' )
 		rotation = 0.0;
 
-	    console.log('rotation', rotation);
-	    
-	    // XYCoords
+	    // Important note: this function only works if start- and end-angle are within
+	    // one whole circle [x,x+2*PI].
+	    // Revelations of more than 2*PI might result in unexpected arcs.
+	    // -> Use the geomutils.wrapMax( angle, 2*PI )
+	    startAngle = geomutils.wrapMax( startAngle, Math.PI*2 );
+	    endAngle = geomutils.wrapMax( endAngle, Math.PI*2 )
+
+	    // Find the start- and end-point on the rotated ellipse
+	    // XYCoords to Vertex (for rotation)
 	    var end : Vertex = new Vertex(VEllipse.utils.polarToCartesian( x, y, radiusH, radiusV, endAngle ));
 	    var start : Vertex = new Vertex(VEllipse.utils.polarToCartesian( x, y, radiusH, radiusV, startAngle ));
 	    end.rotate( rotation, { x:x, y:y } );
 	    start.rotate( rotation, { x:x, y:y } );
-
-	    
+  
 	    var diff : number = endAngle-startAngle;
-
-	    /*
-	      var r2d = 180/Math.PI;
-	      console.log( "startAngle", (r2d*startAngle).toFixed(4),
-	      "endAngle", (r2d*endAngle).toFixed(4),
-	      "diff=" + (r2d*diff).toFixed(4) );
-	    */
 
 	    // Boolean stored as integers (0|1).
 	    var largeArcFlag : number;
-	    // var sweepFlag : numner;
 	    if( diff < 0 ) {
 		largeArcFlag = Math.abs(diff) < Math.PI ? 1 : 0;
-		// sweepFlag = 1;
 	    } else { 
 		largeArcFlag = Math.abs(diff) > Math.PI ? 1 : 0;
-		// sweepFlag = 1;
 	    }
 	    
 	    const sweepFlag : number = 1;
@@ -145,7 +141,9 @@ export class VEllipseSector {
 	    if( options.moveToStart ) {
 		pathData.push('M', start.x, start.y );
 	    }
-	    pathData.push("A", radiusH, radiusV, rotation, largeArcFlag, sweepFlag, end.x, end.y );
+	    // Arc rotation in degrees, not radians.
+	    const r2d : number = 180/Math.PI;
+	    pathData.push("A", radiusH, radiusV, rotation*r2d, largeArcFlag, sweepFlag, end.x, end.y );
 	    return pathData;
 	} // END function describeSVGArc
     } // END ellipseSectorUtils