+++ src/ts/geomutils.ts
@@ -56,14 +56,31 @@ export const geomutils = {
 	return result; 
     },
 
-    // https://stackoverflow.com/questions/4633177/c-how-to-wrap-a-float-to-the-interval-pi-pi
+    /**
+     * Wrap the value (e.g. an angle) into the given range of [0,max).
+     *
+     * @name wrapMax
+     * @param {number} x - The value to wrap.
+     * @param {number} max - The max bound to use for the range.
+     * @return {number} The wrapped value inside the range [0,max).
+     */
     wrapMax( x:number, max:number ) : number {
+	// Found at
+	//    https://stackoverflow.com/questions/4633177/c-how-to-wrap-a-float-to-the-interval-pi-pi
 	return (max + (x%max)) % max;
-    }
-    /* Currently unused
-    wrapMinMax( x:number, min:number, max:number ) : number {		
+    },
+    
+    /**
+     * Wrap the value (e.g. an angle) into the given range of [min,max).
+     *
+     * @name wrapMinMax
+     * @param {number} x - The value to wrap.
+     * @param {number} min - The min bound to use for the range.
+     * @param {number} max - The max bound to use for the range.
+     * @return {number} The wrapped value inside the range [min,max).
+     */
+    // Currently un-used
+    /* wrapMinMax( x:number, min:number, max:number ) : number {		
 	return min + wrapMax( x - min, max -min );
-    }
-    */
-
+    } */
 };